Cache compiled breakpoint condition expressions
Encapsulate expressions in new Expression class that stores the compiled expression and associated variable list.
