+++ Source/Core/Core/CMakeLists.txt
@@ -424,6 +424,8 @@ add_library(core
   PowerPC/BreakPoints.h
   PowerPC/ConditionRegister.cpp
   PowerPC/ConditionRegister.h
+  PowerPC/Expression.cpp
+  PowerPC/Expression.h
   PowerPC/JitInterface.cpp
   PowerPC/JitInterface.h
   PowerPC/MMU.cpp
